+++ px/util.py
@@ -1,6 +1,7 @@
 import os
 import pwd
 
+
 def prompt_yes_no(question):
   while "the answer is invalid":
       reply = str(input(question+' (y/n): ')).lower().strip()
@@ -9,5 +10,12 @@ def prompt_yes_no(question):
       if reply[:1] == 'n':
           return False
 
-def get_user_name():
-  return pwd.getpwuid( os.getuid() )[ 0 ]
+
+def get_user():
+  username = pwd.getpwuid( os.getuid() )[ 0 ]
+  is_root = False
+
+  if username == 'root':
+    is_root = True
+
+  return username, is_root