+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
+## [0.0.2]
+### Added
+
+- Support to apply Flatpak `--user` related updates
+
+### Fixes
+
+- Refactored application into Flatpak and Guix class
+- Properly handle missing parameters
+
 ## [0.0.1]
 ### Added